refactor: #86 리팩토링으로 인한 import 경로명 및 불필요한 파일 제거
